Created getIsSupervisorAttribute in Users Modal.
top-menu.blade.php has a condition where if the user is not a supervisor or admin, ver FTEs and "Supervisión" will not display and if the user is not admin, "Configuración" will not display either.

In ApproveInformes.vue added new actions in the "acciones" button and the full name of the creator of each report is displayed.

In InformeController.php remove the "->latest()" function found in the $plan_id variable because it caused the user's reports to display in ApproveInformes.vue instead of InformesIndex.vue