Changed the home blade to display the apps differently depending on the user's resolution.
Notifications DO NOT send emails to users whenever a notification is called.

When a user leaves an empty field in the contacts section of a report, the application saves the empty fields as 0 instead of null.

Users may solicit access to their report if report is in status "PENDIENTE A ENMIENDA".