Requests that were approved will close on the 15th of each month automatically, changes can be found in app.js.
In ApproveRequestedMonths.vue when it is the 15th of any month, supervisors will not be able to approve or reject requests until the next day.

Reactivated the notifications in PlansController, InformeController.php, RequestedMonthsController.php and UnplannedInformeController.php (the email messages remain deactivated, the reject notifications will now display the rejection reason). New changes on who will recieve notifications using the function NotifySupervisors found in the controllers mentioned before.

New message explaining the create request modal found in Informes.vue. When the pop up for requests submitted, a link will appear in the bottom to redirect the user to the component UserRequestedMonths.vue.

New scopes and attributes in User.php (these are used in NotifySupervisors).